mv some files
